// src/main/requests/types/UpdateBuildingFacilitiesRequest.ts
export interface UpdateBuildingFacilitiesRequest {
  parking: boolean;
  freeWaterDispensers: boolean;
  snackVendingMachines: boolean;
  drinkVendingMachines: boolean;
  cafeteria: boolean;
  waitingArea: boolean;
  waitingAreaChildren: boolean;
  quietRoom: boolean;
  babyChanging: boolean;
  wifi: boolean;
}